requirements.txt
discordpy
Pillow
pymongo
aiodns